refactor(direct-control): extract settlement recommendation source
Move the settlement recommendation embedded source into packages/civ7-direct-control/src/play/tactical/settlement.ts while keeping the public wrapper and command builder in the package facade. Duplicate the tiny probe helper locally to avoid introducing a broad runtime-source helper module.

Record the read-only atom report disposition in the support modularization OpenSpec workstream. This is a source relocation with unchanged fake-tuner package proof, not runtime proof.

Validation: bun run --cwd packages/civ7-direct-control test -- settlement-recommendations.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
